fixed proposal actions

+++ src/app/api/employees/proposals/[id]/route.ts
@@ -29,6 +29,12 @@ export async function PUT(
 
     const { accept, decline } = await req.json();
 
+    if (accept && decline)
+      return NextResponse.json(
+        { message: "Something went wrong" },
+        { status: 400 }
+      );
+
     if (decline) {
       await proposal.deleteOne();
